Fixed issue when the projectile shooter is not player
- Improved game leave
- Temporary removed event when you have game in another world, will check that and this caused player left from game
- Fixed issue when not enough players are in game then other players can't leave from game anymore
- Improved release downloader when we have other double format version

+++ src/main/java/hu/montlikadani/RageMode/utils/UpdateDownloader.java
@@ -27,10 +27,11 @@ public class UpdateDownloader {
 		}
 
 		String msg = "";
+		String versionString = "";
 		String lineWithVersion = "";
 
-		double newVersion = 0d;
-		double currentVersion = 0d;
+		int newVersion = 0;
+		int currentVersion = 0;
 
 		try {
 			URL githubUrl = new URL(
@@ -45,15 +46,12 @@ public class UpdateDownloader {
 				}
 			}
 
-			String[] nVersion;
-			String[] cVersion;
+			versionString = lineWithVersion.split(": ")[1];
+			String nVersion = versionString.replaceAll("[^0-9]", "");
+			newVersion = Integer.parseInt(nVersion);
 
-			String versionString = lineWithVersion.split(": ")[1];
-			nVersion = versionString.replaceAll("[^0-9.]", "").split("\\.");
-			newVersion = Double.parseDouble(nVersion[0] + "." + nVersion[1]);
-
-			cVersion = RageMode.getInstance().getDescription().getVersion().replaceAll("[^0-9.]", "").split("\\.");
-			currentVersion = Double.parseDouble(cVersion[0] + "." + cVersion[1]);
+			String cVersion = RageMode.getInstance().getDescription().getVersion().replaceAll("[^0-9]", "");
+			currentVersion = Integer.parseInt(cVersion);
 
 			if (newVersion > currentVersion) {
 				if ("player".equals(sender)) {
@@ -97,6 +95,12 @@ public class UpdateDownloader {
 						}
 
 						File jar = new File(updatesFolder + per + name + ".jar");
+						if (jar.exists()) {
+							in.close();
+							cancel();
+							return;
+						}
+
 						Files.copy(in, jar.toPath(), StandardCopyOption.REPLACE_EXISTING);
 
 						in.close();
